@@ -56,13 +55,13 @@
     "npm-run-all": "^4.1.5",
     "object-assign": "4.1.1",
     "postcss-flexbugs-fixes": "3.2.0",
-    "postcss-loader": "2.0.8",
+    "postcss-loader": "^8.1.1",
     "promise": "8.0.1",
     "qs": "^6.5.1",
     "raf": "3.4.0",
     "rambda": "^1.1.0",
     "ramda": "^0.24.1",
-    "react": "^16.9.0",
+    "react": "^18.3.1",
     "react-big-calendar": "^0.38.9",
     "react-chartjs-2": "4.0.0",
     "react-color": "^2.14.1",
@@ -157,22 +155,24 @@
     ]
   },
   "devDependencies": {
-    "@babel/plugin-proposal-class-properties": "^7.18.6",
-    "@babel/plugin-transform-nullish-coalescing-operator": "^7.24.1",
-    "@babel/plugin-transform-optional-chaining": "^7.24.5",
-    "@babel/preset-react": "^7.24.1",
-    "@babel/preset-typescript": "^7.24.1",
+    "@babel/core": "^7.24.6",
+    "@babel/preset-env": "^7.24.6",
+    "@babel/preset-react": "^7.24.6",
     "@sentry/webpack-plugin": "1.18.5",
+    "cssnano": "^7.0.1",
     "eslint-loader": "^4.0.2",
     "extract-text-webpack-plugin": "^4.0.0-beta.0",
     "file-loader": "^6.2.0",
-    "html-webpack-plugin": "^4.5.2",
+    "html-webpack-plugin": "^5.5.0",
     "loader-utils": "^1.2.3",
+    "mini-css-extract-plugin": "^2.5.2",
     "npm-run-all": "^4.1.5",
+    "postcss": "^8.4.5",
     "react-dev-utils": "^6.1.1",
-    "terser-webpack-plugin": "^4.2.3",
-    "webpack": "^4.16.5",
-    "webpack-cli": "^4.6.0",
+    "terser-webpack-plugin": "^5.3.10",
+    "webpack": "^5.91.0",
+    "webpack-cli": "^5.1.4",
+    "webpack-dev-server": "^5.0.4",
     "webpack-manifest-plugin": "^2.2.0"
   },
   "eslintConfig": {
