@@ -1,11 +1,13 @@
-import React from 'react'
-import ReactDOM from 'react-dom'
-import { App } from './App'
+import React from "react";
+import { createRoot } from "react-dom/client";
+import { App } from "./App";
+import { HMSRoomProvider } from "@100mslive/react-sdk";
 
-
-  ReactDOM.render(
-    <React.StrictMode>
+const root = createRoot(document.getElementById("root"));
+root.render(
+  <React.StrictMode>
+    <HMSRoomProvider>
       <App />
-    </React.StrictMode>,
-    document.getElementById('root')
-  );
+    </HMSRoomProvider>
+  </React.StrictMode>
+);
