@@ -1,8 +1,11 @@
 import React from 'react';
-import { Tldraw } from '@tldraw/tldraw';
-import '@tldraw/tldraw/tldraw.css'
+import { Whiteboard } from '@100mslive/hms-whiteboard';
+import {  useWhiteboard } from '@100mslive/react-sdk';
+import '@100mslive/hms-whiteboard/index.css';
 export const App = () => {
+  const { token, endpoint, zoomToContent } = useWhiteboard();
+
   return (
-    <Tldraw />
+    <Whiteboard token={token} endpoint={`https://${endpoint}`} zoomToContent={zoomToContent} />
   );
 };
